+++ src/schemas/responses/auth.response.ts
@@ -1,21 +1,12 @@
 import { z } from '../../libs/zod.js';
 
 import { SuccessResponseSchema } from "../responses.js";
-import { UUIDSchema } from '../common.js';
+import {
+  ClientProfileObjectSchema,
+  WorkerProfileObjectSchema,
+  UserObjectSchema,
+} from '../common.js';
 
-
-export const UserObjectSchema = z.object({
-  id: UUIDSchema,
-  phoneNumber: z.string(),
-  firstName: z.string(),
-  lastName: z.string(),
-  middleName: z.string().nullable().optional(),
-  email: z.string().nullable().optional(),
-  status: z.enum(['ACTIVE', 'SUSPENDED', 'BANNED']),
-  isOnline: z.boolean(),
-  createdAt: z.date(),
-  updatedAt: z.date(),
-});
 
 export const RequestOTPResponseSchema = SuccessResponseSchema(z.object({
   phoneNumber: z.string().describe("e.g. '+201234567890'"),
@@ -28,8 +19,8 @@ export type RequestOTPResponseDTO = z.infer<typeof RequestOTPResponseSchema>;
 export const VerifyOTPResponseSchema = SuccessResponseSchema(z.object({
   tokenType: z.enum(['login', 'register']),
   token: z.string().describe("e.g. 'eyJhbGciOiJI...'"),
-  isWorker: z.boolean().optional(),
-  isWorkerSignedUp: z.boolean().optional(),
+  isWorker: z.boolean(),
+  isWorkerSignedUp: z.boolean(),
 }),);
 export type VerifyOTPResponseDTO = z.infer<typeof VerifyOTPResponseSchema>;
 
@@ -53,15 +44,15 @@ export const AccessTokenResponseSchema = SuccessResponseSchema(z.object({
 export type AccessTokenResponseDTO = z.infer<typeof AccessTokenResponseSchema>;
 
 export const ReviewStatusResponseSchema = SuccessResponseSchema(z.object({
-  reason: z.string().optional(),
-  status: z.enum(['PENDING', 'APPROVED', 'REJECTED'], `status must be one of: ${['PENDING', 'APPROVED', 'REJECTED'].join(', ')}`).optional(),
+  reason: z.string(),
+  status: z.enum(['PENDING', 'APPROVED', 'REJECTED'], `status must be one of: ${['PENDING', 'APPROVED', 'REJECTED'].join(', ')}`),
 }),);
 export type ReviewStatusResponseDTO = z.infer<typeof ReviewStatusResponseSchema>;
 
 export const RegisterClientResponseSchema = SuccessResponseSchema(
   z.object({
     user: UserObjectSchema,
-    clientProfile: z.any().describe('ClientProfile object'),
+    clientProfile: ClientProfileObjectSchema.describe('ClientProfile object'),
     accessToken: z.string(),
     refreshToken: z.string(),
   })
@@ -71,7 +62,7 @@ export type RegisterClientResponseDTO = z.infer<typeof RegisterClientResponseSch
 export const RegisterWorkerResponseSchema = SuccessResponseSchema(
   z.object({
     user: UserObjectSchema,
-    workerProfile: z.any().describe('WorkerProfile object'),
+    workerProfile: WorkerProfileObjectSchema.describe('WorkerProfile object'),
     accessToken: z.string(),
     refreshToken: z.string(),
   })